fix: export a browser bundle for generating css
