+++ packages/cli/package.json
@@ -9,13 +9,25 @@
 	"module": "./dist/index.mjs",
 	"types": "./dist/index.d.cts",
 	"exports": {
-		"require": {
-			"types": "./dist/index.d.cts",
-			"default": "./dist/index.cjs"
+		"node": {
+			"require": {
+				"types": "./dist/index.d.cts",
+				"default": "./dist/index.cjs"
+			},
+			"import": {
+				"types": "./dist/index.d.mts",
+				"default": "./dist/index.mjs"
+			}
 		},
-		"import": {
-			"types": "./dist/index.d.mts",
-			"default": "./dist/index.mjs"
+		"default": {
+			"require": {
+				"types": "./dist/browser.d.cts",
+				"default": "./dist/browser.cjs"
+			},
+			"import": {
+				"types": "./dist/browser.d.mts",
+				"default": "./dist/browser.mjs"
+			}
 		}
 	},
 	"scripts": {
