+++ packages/cli/src/index.ts
@@ -3,15 +3,17 @@ export { verify } from './custom/verify';
 export {
 	generateIconStaticCSS,
 	generateIconVariableCSS,
+	generateV1CSS,
+	generateV2CSS,
+	generateVariableCSS,
+} from './google/css';
+export {
 	packagerIconsStatic,
 	packagerIconsVariable,
 } from './google/packager-icons';
-export { generateV1CSS, packagerV1 } from './google/packager-v1';
-export { generateV2CSS, packagerV2 } from './google/packager-v2';
-export {
-	generateVariableCSS,
-	packagerVariable,
-} from './google/packager-variable';
+export { packagerV1 } from './google/packager-v1';
+export { packagerV2 } from './google/packager-v2';
+export { packagerVariable } from './google/packager-variable';
 export { processGoogle } from './google/queue';
 export {
 	APIIconStatic,
